Added the base files for the start of the project
I initally started by adding the base files, index, style, and the logic
for the etch-a-sketch. In index.html, the base layout was planned out
first, then style.css was filled out with basic boilerplate stuff. Then,
I implemented the very basic logic and proof-of-concept for the actual
etch-a-sketch using javascript.